# auth-service/openapi.yaml
openapi: 3.1.0
info:
  title: Auth Service Management API
  version: 1.0.0
  description: Management API for the Goaldone Auth Service. Provides endpoints for user, organization, and invitation management. All endpoints require OAuth2 Bearer token authentication.
  contact:
    name: Goaldone Team
    url: https://goaldone.de
  license:
    name: Proprietary
servers:
  - url: http://localhost:9000
    description: Local development server
  - url: https://auth.goaldone.de
    description: Production server
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from the OAuth2 authorization server
  schemas:
...
    InvitationRequest:
      type: object
      title: InvitationRequest
      description: Request body for creating an invitation
      properties:
        email:
          type: string
          format: email
          description: Email address to invite
          example: newuser@example.com
        companyId:
          type: string
          format: uuid
          description: Organization UUID to invite the user to
          example: 550e8400-e29b-41d4-a716-446655440002
        inviterId:
          type: string
          format: uuid
          description: UUID of the user sending the invitation
          example: 550e8400-e29b-41d4-a716-446655440000
      required:
        - email
        - companyId
        - inviterId
    InvitationResponse:
      type: object
      title: InvitationResponse
      description: Response containing invitation information
      properties:
        id:
          type: string
          format: uuid
          description: Unique invitation identifier
          example: 550e8400-e29b-41d4-a716-446655440003
        email:
          type: string
          format: email
          description: Email address invited
          example: newuser@example.com
        companyId:
          type: string
          format: uuid
          description: Organization UUID
          example: 550e8400-e29b-41d4-a716-446655440002
        inviterId:
          type: string
          format: uuid
          description: UUID of the user who sent the invitation
          example: 550e8400-e29b-41d4-a716-446655440000
        expiresAt:
          type: string
          format: date-time
          description: Timestamp when the invitation expires
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the invitation was created
    InvitationStatusResponse:
      type: object
      title: InvitationStatusResponse
      description: Response containing invitation status information
      properties:
        token:
          type: string
          description: Invitation token
          example: 550e8400-e29b-41d4-a716-446655440004
        status:
          type: string
          enum: [PENDING, ACCEPTED, DECLINED, EXPIRED, NOT_FOUND]
          description: Invitation status
          example: PENDING
        invitedEmail:
          type: string
          format: email
          description: Email address that was invited
          example: newuser@example.com
        organizationId:
          type: string
          format: uuid
          description: Organization UUID
          example: 550e8400-e29b-41d4-a716-446655440002
        organizationName:
          type: string
          description: Organization name
          example: Acme Corporation
        expirationDate:
          type: string
          format: date-time
          description: Timestamp when the invitation expires
        emailMatch:
          $ref: '#/components/schemas/EmailMatch'
    EmailMatch:
      type: object
      title: EmailMatch
      description: Email matching result
      properties:
        found:
          type: boolean
          description: Whether a user with this email already exists
          example: false
        userId:
          type: string
          format: uuid
          description: UUID of the user if a match was found
          example: 550e8400-e29b-41d4-a716-446655440000
    ProblemDetail:
      type: object
      description: HTTP problem detail response (RFC 7807)
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type
        title:
          type: string
          description: A short, human-readable summary of the problem
        status:
          type: integer
          description: The HTTP status code
        detail:
          type: string
          description: A human-readable explanation of the problem
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence
paths:
...
  /api/v1/invitations:
    post:
      tags:
        - Invitations
      summary: Create a new invitation
      description: Creates a new invitation for a user to join an organization
      operationId: createInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationRequest'
      responses:
        '201':
          description: Invitation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
        '409':
          description: Invitation already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/invitations/{token}:
    get:
      tags:
        - Invitations
      summary: Get invitation by token
      description: Retrieves invitation details by its unique token
      operationId: getInvitationByToken
      parameters:
        - name: token
          in: path
          required: true
          description: Invitation token
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invitation found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResponse'
        '401':
          description: Unauthorized
        '404':
          description: Invitation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '410':
          description: Invitation has expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    delete:
      tags:
        - Invitations
      summary: Cancel invitation
      description: Cancels an existing invitation
      operationId: cancelInvitation
      parameters:
        - name: token
          in: path
          required: true
          description: Invitation token
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Invitation cancelled successfully
        '401':
          description: Unauthorized
        '404':
          description: Invitation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/invitations/{token}/status:
    get:
      tags:
        - Invitations
      summary: Get invitation status
      description: Queries the status of an invitation using its token. This endpoint is public and does not require authentication.
      operationId: getInvitationStatus
      security: []
      parameters:
        - name: token
          in: path
          required: true
          description: Invitation token
          schema:
            type: string
      responses:
        '200':
          description: Invitation found and is pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationStatusResponse'
        '404':
          description: Invitation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationStatusResponse'
        '409':
          description: Invitation has already been accepted or declined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationStatusResponse'
        '410':
          description: Invitation has expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationStatusResponse'
...
  /api/v1/admin/users/{userId}/super-admin-status:
    patch:
      tags:
        - Admin
      summary: Update user super-admin status
      description: Grants or revokes super-admin status for a user. Cannot remove super-admin status from the last system administrator.
      operationId: updateSuperAdminStatus
      parameters:
        - name: userId
          in: path
          required: true
          description: User UUID
          schema:
            type: string
            format: uuid
        - name: newSuperAdminStatus
          in: query
          required: true
          description: New super-admin status (true to grant, false to revoke)
          schema:
            type: boolean
      responses:
        '200':
          description: Super-admin status updated successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient privileges
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Cannot remove super-admin status from the last system administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'